fix: 修复 Window ACK 永久阻塞及 UInt32 溢出问题
- WindowControl 新增 ACK 超时自动禁用机制，服务器不发 ACK 时 5 秒后自动跳过流控，收到 ACK 后自动恢复
- 计数器类型 UInt32 → UInt64，防止 10Mbps 码率下约 55 分钟溢出
- 修复窗口乘法溢出：windowSize * seq 改为 UInt64(windowSize) * seq
- TransmissionStatistics 同步更新 totalBytesReceived/Sent 为 UInt64
- 新增 3 个测试：ACK 超时禁用、ACK 恢复重启、UInt64 无溢出

+++ Sources/HPRTMP/Core/TransmissionStatistics.swift
@@ -10,8 +10,8 @@ import Foundation
 public struct TransmissionStatistics: Sendable {
   public let pendingMessageCount: Int
 
-  public let totalBytesReceived: UInt32
-  public let totalBytesSent: UInt32
+  public let totalBytesReceived: UInt64
+  public let totalBytesSent: UInt64
   public let unacknowledgedBytes: Int64
   public let windowSize: UInt32
   public let windowUtilization: Double
